pyproject.toml
[project]
name = "nl2sql-v3"
version = "0.1.0"
description = "NL2SQL v3 - Multi-agent SQL generation system with LangGraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # LangGraph核心
    "fastapi>=0.120.3",
    "langchain>=1.0.0",
    "langgraph>=1.0.0",
    "uvicorn>=0.38.0",
    # 数据库驱动
    "psycopg[binary]>=3.1.0",
    "pgvector>=0.2.0",
    "neo4j>=5.15.0",
    # SQL处理和验证
    "sqlparse>=0.4.4",
    "pydantic>=2.5.2",
    # LLM和嵌入
    "dashscope>=1.14.0",
    "langchain-openai>=0.2.0",
    # 工具
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.0",
    "langchain-community>=0.4.1",
    "psycopg-pool>=3.2.7",
    "pytest>=8.4.2",
    # MetaWeave 依赖
    "pandas>=2.1.3",
    "tqdm>=4.66.1",
    "click>=8.1.7",
    "jinja2>=3.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
monitoring = [
    "langsmith>=0.1.0",
    "redis>=5.0.0",
]
